hammer_api/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==4.2.23
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
phonenumbers==9.0.10
psycopg2-binary==2.9.3
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
